move tranistion stuff to redshift-transition
